requirements.txt
absl-py==2.1.0
asgiref==3.7.2
attrs==23.2.0
backports.zoneinfo==0.2.1
cffi==1.16.0
contourpy==1.1.1
cvzone==1.6.1
cycler==0.12.1
Django==4.2.9
flatbuffers==23.5.26
fonttools==4.47.2
gunicorn==21.2.0
importlib-resources==6.1.1
kiwisolver==1.4.5
matplotlib==3.7.4
mediapipe==0.10.9
numpy==1.24.4
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
packaging==23.2
pillow==10.2.0
protobuf==3.20.3
pycparser==2.21
pyparsing==3.1.1
python-dateutil==2.8.2
six==1.16.0
sounddevice==0.4.6
sqlparse==0.4.4
typing-extensions==4.9.0
tzdata==2023.4
zipp==3.17.0
